whoops, long, not long long

+++ src/linux/last-rites.c
@@ -112,7 +112,7 @@ static void trim (char *s)
 
 static void sleep(int n)
 {
-    long long s[2] = { n, 0 };
+    long s[2] = { n, 0 };
 
     sys_select (0, (void*)0, (void*)0, (void*)0, s);
 }
